Revert "Modal"
