--- src/App.jsx
@@ -1,18 +1,0 @@
-import Home from "./pages/Home";
-
-import Footer from "./components/Footer";
-import Header from "./components/Header";
-
-import "./app.scss";
-
-const App = () => {
-  return (
-    <>
-      <Header/>
-      <Home />
-      <Footer/>
-    </>
-  );
-};
-
-export default App;
