+++ src/app.scss
@@ -1,6 +1,6 @@
-@import './styles/variables/variables.scss';
-@import './styles/normalize/normalize.scss';
-@import './styles/font-family/font-family.scss';
+@import './stylesUniversal/variables/variables.scss';
+@import './stylesUniversal/normalize/normalize.scss';
+@import './stylesUniversal/font-family/font-family.scss';
 
 * {
   box-sizing: border-box;
