--- src/components/Button/Button.jsx
@@ -1,10 +1,0 @@
-import "./button.scss";
-import clsx from 'clsx';
-
-const Button = ({ text, className, onClick }) => {
-  return (<button className={clsx(className)} onClick={onClick}>
-    {text}
-  </button>);
-};
-
-export default Button;
