+++ src/componentsBasic/home/homeSections/homeSubscribe/HomeSubscribe.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/variables/variables.scss';
+@import '../../../../stylesUniversal/variables/variables.scss';
 
 .subscribe {
   font-family: "Poppins-Regular", "sans-serif";
