+++ src/componentsUniversal/button/button.js
@@ -1,0 +1,9 @@
+import './Button.scss'
+
+const Button = ({text, className=''}) => {
+    return (
+        <button  className={`btn ${className}`}>{text}</button>
+    )
+}
+
+export default Button;
