+++ src/componentsUniversal/button/button.scss
@@ -1,5 +1,5 @@
-@import '../../styles/mixins/mixins.scss';
-@import '../../styles/variables/variables.scss';
+@import '../../stylesUniversal/mixins/mixins.scss';
+@import '../../stylesUniversal/variables/variables.scss';
 
 .btn {
     @include BTN ($white, 600, $main-color, none, 10px, flex, row, center, center);
