+++ src/componentsUniversal/footer/footer.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables/variables.scss';
+@import '../../stylesUniversal/variables/variables.scss';
 
 .footer {
   font-family: "Poppins-Light", "sans-serif";
