+++ src/componentsUniversal/header/header.scss
@@ -1,4 +1,4 @@
-@import '../../styles/mixins/mixins.scss';
+@import '../../stylesUniversal/mixins/mixins.scss';
 
 .header {
   font-family: 'Baloo2-Regular', sans-serif;
