--- src/pages/Home/Home.jsx
@@ -1,13 +1,0 @@
-import Info from "./HomeSections/HomeInfo";
-import Subscribe from "./HomeSections/HomeSubscribe";
-
-const Home = () => {
-  return (
-    <>
-      <Info />
-      <Subscribe />
-    </>
-  );
-};
-
-export default Home;
