--- src/pages/Home/HomeSections/HomeInfo/index.js
@@ -1,1 +1,0 @@
-export {default} from './HomeInfo';
