--- src/pages/Home/HomeSections/HomeSubscribe/index.js
@@ -1,1 +1,0 @@
-export {default} from './HomeSubscribe';
