--- src/pages/Modal/Modal.scss
@@ -1,77 +1,0 @@
-@import '../../styles/variables/variables.scss';
-
-.modal {
-  position: fixed;
-  left: 0;
-  right: 0;
-  top: 0;
-  bottom: 0;
-  background-color: inherit;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: $modal-background;
-  z-index: 5;
-  &__content {
-    width: 776px;
-    height: 512px;
-    background-color: $white;
-    z-index: 10;
-    display: flex;
-    justify-content: space-between;
-    border-radius: 15px;
-  }
-  &__img {
-    width: 260px;
-    border-radius: 0px 15px 15px 0px;
-  }
-  &__main {
-    display: flex;
-    flex-direction: column;
-    padding: 32px 32px 36px 32px;
-  }
-  &__header-close {
-    width: 16px;
-    height: 16px;
-    margin-bottom: 124px;
-    &:hover {
-      cursor: pointer;
-    }
-  }
-  &__body {
-    &-text {
-      &-form {
-        display: flex;
-      }
-      &-email {
-        width: 220px;
-        height: 40px;
-        margin-right: 24px;
-        border-radius: 10px;
-        border: 1px $input-border solid;
-        padding: 10px 0 10px 16px;
-        &:hover, :active, :focus {
-            border: 1px $input-border-hover solid;
-        }
-      }
-      &-title {
-        font-family: 'Baloo2-Regular', sans-serif;
-        font-size: 28px;
-        line-height: 36px;
-        font-weight: 400;
-        margin-bottom: 20px;
-      }
-    }
-    &-btn {
-      width: 152px;
-      height: 40px;
-    }
-  }
-  &__footer-text {
-    font-family: 'Poppins-Light', sans-serif;
-    font-size: 10px;
-    line-height: 12px;
-    text-align: center;
-    margin-top: 182px;
-  }
-}
