--- src/styles/variables/variables.scss
@@ -1,9 +1,0 @@
-$white: #FFFFFF;
-$black: #000000;
-$main-color: #59abb0;
-$button-hover: #1B6C71;
-
-$input-border: #C2D2D2;
-$input-border-hover: $button-hover;
-
-$modal-background: rgb(0, 0, 0, 0.5)
