+++ src/stylesUniversal/variables/variables.scss
@@ -1,0 +1,4 @@
+$white: #FFFFFF;
+$black: #000000;
+$main-color: #59abb0;
+$button-hover: #1B6C71;
